+++ runtime/service/src/testFixtures/java/org/apache/polaris/service/TestServices.java
@@ -303,8 +303,7 @@ public record TestServices(
               vendingMechanisms,
               () -> GoogleCredentials.create(new AccessToken(GCP_ACCESS_TOKEN, new Date())),
               storageCredentialCache,
-              realmConfig,
-              diagnostics);
+              realmConfig);
       InMemoryPolarisMetaStoreManagerFactory metaStoreManagerFactory =
           new InMemoryPolarisMetaStoreManagerFactory(
               clock, diagnostics, storageIntegrationProvider, RootCredentialsSet.EMPTY);